shewabingo-bot/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
eventlet==0.36.1
python-telegram-bot==21.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
Pillow==10.4.0
gunicorn==21.2.0
python-dotenv==1.0.1
